requirements.txt
# LangGraph Orchestrator Dependencies

# Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0

# HTTP Client (async)
httpx>=0.27.0

# Database (für Checkpointing)
psycopg2-binary>=2.9.0
# oder: asyncpg>=0.29.0 für async

# Optional: Tavily für Web-Suche
tavily-python>=0.5.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
